dep updates - cleaning up the helper function for doing the pdf compares
